Refactor signature property tests and stored file handling
Removed outdated cassette file and refactored signature property tests into a new module. Updated stored file handling to support in-memory image files for uploads. Minor changes to property reference method to access stored files by ID.
