+++ tests/test_properties.py
@@ -1,16 +1,13 @@
-import os
 from datetime import date, datetime, time
-from unittest import skip
 
 from pykechain.enums import (
     Category,
     ContextGroup,
     Multiplicity,
     PropertyType,
-    StoredFileClassification,
 )
 from pykechain.exceptions import APIError, IllegalArgumentError, NotFoundError
-from pykechain.models import Property, SignatureProperty
+from pykechain.models import Property
 from pykechain.models.validators import SingleReferenceValidator
 from tests.classes import TestBetamax
 
@@ -562,58 +559,3 @@ class TestPropertiesWithReferenceProperty(TestBetamax):
 
         # tearDown
         copied_ref_property.delete()
-
-
-class TestPropertiesSignatureProperty(TestBetamax):
-    def setUp(self):
-        super().setUp()
-
-        self.wheel_model = self.project.model("Wheel")
-        self.bike = self.project.model("Bike")
-
-        self.test_files_dir = os.path.dirname(
-            os.path.dirname(os.path.abspath(__file__)).replace("\\", "/")
-        )
-        self.upload_path = os.path.join(
-            self.test_files_dir,
-            "tests",
-            "files",
-            "test_upload_image_to_attachment_property",
-            "test_upload_image.png",
-        )
-        self.file_name = "__TEST_SIGNATURE_STORED_FILE"
-        self.signature_stored_file = self.client.create_stored_file(
-            name=self.file_name,
-            scope=self.project,
-            classification=StoredFileClassification.SCOPED,
-            filepath=self.upload_path,
-            description="__TEST_STORED_FILE_DESCRIPTION",
-        )
-
-        self.test_signature_property_model: SignatureProperty = self.bike.add_property(
-            name=f"__Test signature property @ {datetime.now()}",
-            property_type=PropertyType.SIGNATURE_VALUE,
-            description="Description of test signature property",
-            unit="no unit",
-            default_value=None,
-        )
-
-    def tearDown(self):
-        self.test_signature_property_model.delete()
-        self.signature_stored_file.delete()
-        super().tearDown()
-
-    def test_create_signature_property(self):
-        # implicit inside the setUp
-        pass
-
-    @skip("TODO: fix this test")
-    def test_upload_new_signature_to_property(self):
-        self.test_signature_property_model.value = self.signature_stored_file
-        ### Results in a 400
-        # [
-        #     "Cannot assign \"'151b1665-d738-40a3-a699-be32690d88b2'\": \"Property.signature_val\" must be a \"StoredFile\" instance."
-        # ]
-
-        self.test_signature_property_model.refresh()
-        self.assertIsNotNone(self.test_signature_property_model.value)
