+++ tests/utils.py
@@ -1,3 +1,6 @@
+from io import BytesIO
+
+from PIL import Image
 from envparse import Env
 
 # reads a local .env file with the TEST_TOKEN=<user token>
@@ -11,3 +14,11 @@ TEST_TOKEN = env("TEST_TOKEN", default="")
 TEST_SCOPE_ID = env("TEST_SCOPE_ID", default="bd5dceaa-a35e-47b0-9fc5-875410f4a56f")
 TEST_SCOPE_NAME = env("TEST_SCOPE_NAME", default="Bike Project")
 TEST_RECORD_CASSETTES = env.bool("TEST_RECORD_CASSETTES", default=True)
+
+
+def create_test_image_file() -> BytesIO:
+    """Create an Image file in memory for testing handling upload of Images."""
+    image_io = BytesIO()
+    Image.new("RGB", (50, 50), color="red").save(image_io, format="jpeg")
+    image_io.seek(0)
+    return image_io
